Several changes, read extended commit message
Fix bugs of features
introduced in 7658bc9f6b06459f22b7e2a09b3bdcab98873081.

More fit variables to rabi result class. Now this fit is
exactly corresponds to the theoretical model.

Minor restyling/renaming of the modules and packages inside lib2.
This leads to better module/package structure of the library.

Fix Spectrum_m4x behaviour when he misses timeout error.

drivers\Spectrum_m4x.py
	fix
		Fixing problem of the previous commit changes
		external trigger was never setted by 'setup_trigger_source'

DigitizerDirectRabi->lib2\digitizerPulsedMeasurments\directRabi.py
	feat
		additional offset variables added to the model function.

style
	Minor changes
	Files rearranging, renaming and corresponding refactoring

lib2/pulseMixing.py
	feat
		Added opportunity to cut out excitation pulses
		from valuable signal.

drivers/Spectrum_m4x.py
	fix
		Manual for Spectrum, p 47:
		... GetErrorInfo function won’t return the timeout event
		even if it had occurred. … You can only recognize the
		ERR_TIMEOUT as a direct return value of the wait function
		that was called.
		Existed ‘self.__handle_error’ could not catch timeout
		Neccessary changes in driver were provided and tested.

+++ drivers/keysightM3202A.py
@@ -286,7 +286,7 @@ class KeysightM3202A(Instrument):
         # duration = duration_initial*1e4 if duration_initial < 1e2 else 1e6  # here it is
 
         duration = duration_initial
-        new_x = np.linspace(0, duration, np.round(duration/self.get_sample_rate()*1e9), endpoint=False)
+        new_x = np.linspace(0, duration, int(np.round(duration/self.get_sample_rate()*1e9)), endpoint=False)
 
         # converting domain values in the function domain
         new_x_converted = np.remainder(new_x, duration_initial)
